worker/requirements.txt
redis==5.2.1
rq==2.0.0
python-dotenv==1.0.1
